requirements.txt
Django==3.1.*
django-concurrency==2.2.*
django-debug-toolbar==3.1.*
django-extensions==3.1.*
psycopg2-binary==2.8.*
